requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
python-fasthtml>=1.0.0
MonsterUI>=0.1.0
